procfile change, settings change to work local
